solved warnings

+++ src/main/fission_simulator.cpp
@@ -128,20 +128,20 @@ int main(int argc, char** argv)
     uint64_t t0 =
             std::chrono::duration_cast<std::chrono::microseconds>
             (std::chrono::system_clock::now().time_since_epoch()).count();
-    for(int i = 0; i < nSim;)
+    for(uint32_t i = 0; i < nSim;)
     {
-        std::vector<int> fissionProducts = fGen.generateFissionValues(distMass, distAtom, massNumbers, atomNumbers);
+        std::vector<double> fissionProducts = fGen.generateFissionValues(distMass, distAtom, massNumbers, atomNumbers);
 
-        int a1 = fissionProducts.at(0);
+        int a1 = static_cast<int>(fissionProducts.at(0));
         countsPerA[a1 - MIN_MASS_NUMBER] += 1;
 
-        int z1 = fissionProducts.at(1);
+        int z1 = static_cast<int>(fissionProducts.at(1));
         countsPerZ[z1 - MIN_ATOMIC_NUMBER] += 1;
 
-        int a2 = fissionProducts.at(2);
+        int a2 = static_cast<int>(fissionProducts.at(2));
         countsPerA[a2 - MIN_MASS_NUMBER] += 1;
 
-        int z2 = fissionProducts.at(3);
+        int z2 = static_cast<int>(fissionProducts.at(3));
         countsPerZ[z2 - MIN_ATOMIC_NUMBER] += 1;
 
         double free = fissionProducts.at(4);
